Auth - API (invert_hash)
